devtools/client/shared/components/reps/test/mochitest/test_reps_failure.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test fallback for rep rendering when a rep fails to render.
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - Failure</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript;version=1.8"></script>
<script type="application/javascript;version=1.8">
window.onload = Task.async(function* () {
  try {
    const { REPS } = browserRequire("devtools/client/shared/components/reps/load-reps");
    let { Rep, ArrayRep, RegExp } = REPS;

    // Force the RegExp rep to crash by creating RegExp grip that throws when accessing
    // the displayString property
    let gripStub = {
      "type": "object",
      "class": "RegExp",
      "actor": "server1.conn22.obj39",
      "extensible": true,
      "frozen": false,
      "sealed": false,
      "ownPropertyLength": 1,
      get displayString() {
        throw new Error("failure");
      }
    };

    // Test that correct rep is chosen.
    const renderedRep = shallowRenderComponent(Rep, { object: gripStub });
    is(renderedRep.type, RegExp.rep, `Rep correctly selects ${RegExp.rep.displayName}`);

    // Test fallback message is displayed when rendering bad rep directly.
    let renderedComponent = renderComponent(RegExp.rep, { object: gripStub });
    is(renderedComponent.textContent, "Invalid object", "Fallback rendering has expected text content");

    // Test fallback message is displayed when bad rep is nested in another rep.
    renderedComponent = renderComponent(ArrayRep.rep, { object: [1, gripStub, 2] });
    is(renderedComponent.textContent, "[ 1, Invalid object, 2 ]", "Fallback rendering has expected text content");
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
